Create Lesson 4 Birthday Card
Created a birthday card app with a relative layout. Happy Birthday Paul is in the top left corner, From Matt is in the bottom right. They are in a relative layout on top of the image of androids eating cake, which was cropped to cover the whole screen with the correct pixel size.